Fix markers not being added on a big zoom animation. Fixes #216
If we zoomed down such that a cluster would be out of our visible bounds then any of its children would be added at its location for the animation. In this case those markers would immediately be removed as they are out of our visible bounds.
Instead we now don't animate adding markers after a zoom if they are coming from outside of our visible bounds.