Added animate option
Following issues #382, #348 and PR #352, added a new option `animate` to
`L.MarkerClusterGroup` to select the animation behaviour. It defaults to
`L.DomUtil.TRANSITION` so that if it is left unassigned, MCG keeps the
current behaviour (it animates or not depending on `L.DomUtil.TRANSITION`)
but it is determined at instanciation time instead of at script loading time.
To achieve this, this commit removes the test condition when including the
animation (or non-animation) methods. It adds all methods to prototype, and
hooks to the appropriate one at initialization.
`L.MarkerCluster` is also adapted, because it animates when spiderfying.
In order to avoid having to hook at each instanciation, this commit actually
creates an extended class `L.MarkerClusterNonAnimated` which includes the
non-animated methods. Then `L.MarkerClusterGroup` instantiates the appropriate
class.