Rebased and improved default behaviour
Rebased to get more reliable test suites.

Changed the default behaviour of `animate` option: defaults to true,
but if `L.DomUtil.TRANSITION` is false, we always use the non-animated
versions, no matter the setting of `animate`, because Leaflet will not
be able to animate anything.

Improved README doc about this option and its effect on `animateAddingMarkers`