Added animate option
Deleted previous branch animationOption since I messed up trying to rebase...

Following issues #382, #348 and PR #352, added a new option `animate` to
`L.MarkerClusterGroup` to select the animation behaviour.

It defaults to true and its effect is ignored if `L.DomUtil.TRANSITION` is
false, so that MCG keeps the current behaviour (it animates or not depending
on `L.DomUtil.TRANSITION`) when this option is left unassigned.

The only difference is that its effect is determined at instanciation time
instead of at script loading time.

To achieve this, this code removes the test condition when including the
animation (or non-animation) methods. It adds all methods to prototype, and
hooks to the appropriate one at initialization.
`L.MarkerCluster` is also adapted, because it animates when spiderfying.
In order to avoid having to hook at each instanciation, this code actually
creates an extended class `L.MarkerClusterNonAnimated` which includes the
non-animated methods. Then `L.MarkerClusterGroup` instantiates the appropriate
class.

Also created a dedicated test spec and included it into spec/index.

Finally updated README to describe this new option and its effect on
`animateAddingMarkers` option.