Corrected spiderfy leg anim and leg opacity option
Spiderfy leg animation should be corrected now for all cases (options set or not, animated or not).

Corrected use of `spiderLegPolylineOptions` so that it defines final opacity in animation mode.
Therefore, we can now specify a default opacity value (chose 0.5 as this was the previous animation final opacity) that drives the animation end.
Updated README in consequence.