Manage adding nested Layer Groups
Following #622 and #623, modified how addLayer manages adding a Layer Group.
It now passes control to addLayers. The latter checks in its main loop if the current layer is a Layer Group.
In that case, it appends all children to its input array and skips the current group.

Also added corresponding test specs in AddLayer.SingleSpec and AddLayersSpec.
Took the opportunity to modify these test suite to avoid PhantomJS memory leak.