Get zoom from internal _zoom instead of map.getZoom()
in addLayer and _animationAddLayer methods, for compatibility with MCG.Freezable plugin (see issue https://github.com/ghybs/Leaflet.MarkerCluster.Freezable/issues/1).
In normal situation (i.e. without Freezable plugin), the 2 should differ only momentarily at the end of a zoom, before the _zoomEnd listener is called, and until that listener completes. If addLayer happens before the listener is called, it should perform as if the zoom did not occur yet, then _zoomEnd proceeds with the new layer / cluster without issue. addLayer should not be able to execute in the middle of _zoomEnd, as there is no asynchronous task involved, except for the end of the zoom animation, where previous clusters are collected. For that one, _zoom is already updated to equal map._zoom, so there is no change compared to before this commit.
When Freezable plugin is added, it tampers _zoom and listeners to disable changes and animations. So we should no longer call map.getZoom() / map._zoom outside the disabled listeners.