Add specs for handling ES6 iterators to addLayers
This is an initial attempt to spec out what would be expected if this
plugin supported ES6 iterators.

An example usage might look like:

    function *mapMarkers() {
      yield new L.Marker([1.5, 1.5]);
      yield new L.Marker([1.5, 1.5]);
      yield new L.Marker([3.0, 1.5]);
    }

    var group = new L.MarkerClusterGroup();
    group.addLayers(mapMarkers());

This (slightly contrived) example could allow the clustering logic to
perform batch processing more efficiently and with less memory over
head.

An example real world case would be when markers are constructed from
Lat/Lng data. In Array based code one has to map over the array to make
a interim Array of `L.Markers` which are passed to `addLayers` which has
to iterate over the array a second time. With the use of an iterator
(typically produced from a generator function) that pipeline could be
reduced considerably by only processing the mapping at the time it is
needed. This means that the interim array described above would not
exists while the `addLayers` batch processed. It also means that the
memory required would only be a large as the number of markers being
processed in any given run loop.

Comparison of a clustering with iterators the
consumption would be at most 1 at a time plus the final set. While
without you would have the interim mapping of markers plus the final
set. This results in `N+1` VS. `N^2`.

This PR is meant as a starting point for discussion (RFC). It currently
is simply a spec around the API usage. Actual implementation *might*
result in a significant rewrite or separate branch of logic. The pro and
cons needs to be clearly discussed before determining if an
implementation of this feature is worth consideration.