Make DistanceGrid cells closed sets
If the closest object in a DistanceGrid lies exactly on the border of a
cell, still return it in getNearObject(). In mathematical terms, this
makes the cell a closed set (that is, a set which contains all of its
boundary points).

A test with two getNearObject() calls is added. The first expectation
works with the old and the new behavior, the second expectation only
succeeds with the new behavior.