Support DistanceGrid with cell size 0
For extremely small cell sizes, fall back to returning the original
coordinate in _getCoord(), instead of Infinity. Together with the
previous change, this groups objects with identical coordinates into the
same cell and therefore fixes #836 for maxClusterRadius values like 0 or
Number.MIN_VALUE. (However, Number.EPSILON is actually large enough that
dividing normal values for x by it does not yet yield Infinity, so in
that case the bug persists.)

We use isFinite() instead of Number.isFinite() because the latter is not
as widely supported (in particular, it doesn’t work in `jake test`).
Since the argument is the result of Math.floor, it should always be a
number, so the difference should not matter.